+++ src/main/java/com/samcomo/dbz/global/config/SecurityConfig.java
@@ -12,8 +12,8 @@ import com.samcomo.dbz.member.jwt.JwtUtil;
 import com.samcomo.dbz.member.jwt.filter.AccessTokenFilter;
 import com.samcomo.dbz.member.jwt.filter.CustomLoginFilter;
 import com.samcomo.dbz.member.jwt.filter.CustomLogoutFilter;
-import com.samcomo.dbz.member.jwt.filter.FilterMemberExceptionHandler;
-import com.samcomo.dbz.member.jwt.filter.LoginSuccessHandler;
+import com.samcomo.dbz.member.jwt.filter.handler.FilterMemberExceptionHandler;
+import com.samcomo.dbz.member.jwt.filter.handler.LoginSuccessHandler;
 import com.samcomo.dbz.member.model.repository.MemberRepository;
 import com.samcomo.dbz.member.service.impl.Oauth2MemberServiceImpl;
 import jakarta.servlet.http.HttpServletRequest;
